+++ lib/td_dd/data_structures/bulk_update.ex
@@ -3,13 +3,10 @@ defmodule TdDd.DataStructures.BulkUpdate do
   Support for bulk update of data structures.
   """
 
-  import Canada, only: [can?: 2]
-
   alias Codepagex
   alias Ecto.Changeset
   alias Ecto.Multi
   alias TdCache.TaxonomyCache
-  alias TdDd.Auth.Claims
   alias TdDd.DataStructures
   alias TdDd.DataStructures.Audit
   alias TdDd.DataStructures.DataStructure
@@ -18,10 +15,13 @@ defmodule TdDd.DataStructures.BulkUpdate do
   alias TdDd.Search.IndexWorker
   alias TdDfLib.Format
   alias TdDfLib.Templates
+  alias Truedat.Auth.Claims
 
   require Logger
 
   @data_structure_preloads [:system, current_version: :structure_type]
+
+  defdelegate authorize(action, user, params), to: __MODULE__.Policy
 
   def update_all(
         ids,
@@ -220,10 +220,10 @@ defmodule TdDd.DataStructures.BulkUpdate do
             params = %{domain_ids: Enum.map(domains, & &1.id)}
             changeset = DataStructures.update_changeset(claims, structure, params)
 
-            if can?(claims, update_data_structure(changeset)) do
+            if Bodyguard.permit?(DataStructures, :update_data_structure, claims, changeset) do
               {index, check_data_structure(changeset)}
             else
-              {index, {:error, {:update_domain, :forbbiden}}}
+              {index, {:error, {:update_domain, :forbidden}}}
             end
           end
       end)