+++ lib/td_dd_web/controllers/grant_request_approval_controller.ex
@@ -1,8 +1,6 @@
 defmodule TdDdWeb.GrantRequestApprovalController do
   use TdDdWeb, :controller
 
-  import Canada, only: [can?: 2]
-
   alias TdDd.Grants.Requests
 
   action_fallback TdDdWeb.FallbackController
@@ -10,7 +8,7 @@ defmodule TdDdWeb.GrantRequestApprovalController do
   def create(conn, %{"grant_request_id" => id, "approval" => params}) do
     with claims <- conn.assigns[:current_resource],
          request <- Requests.get_grant_request!(id, claims),
-         {:can, true} <- {:can, can?(claims, approve(request))},
+         :ok <- Bodyguard.permit(Requests, :approve, claims, request),
          {:ok, %{approval: approval}} <- Requests.create_approval(claims, request, params) do
       conn
       |> put_status(:created)